Корректировка

+++ src/server/game/Spells/Auras/SpellAuras.cpp
@@ -455,7 +455,7 @@ void Aura::_UnapplyForTarget(Unit * target, Unit * caster, AuraApplication * aur
     ASSERT(auraApp);
 
     ApplicationMap::iterator itr = m_applications.find(target->GetGUID());
-
+    
     // TODO: Figure out why this happens
     if (itr == m_applications.end())
     {
@@ -1680,8 +1680,8 @@ void Aura::HandleAuraSpecificMods(AuraApplication const * aurApp, Unit * caster,
             {
                 case 44544: // Fingers of Frost
                 {
-                    int32 key = int32(uint32(uint32(aurApp->GetBase()->GetApplyTime()) & uint32(0x7FFFFFFF)));
- 
+                    int32 key = (aurApp->GetBase()->GetApplyTime() & 0x7FFFFFFF);
+
                     // See if we already have the indicator aura. If not, create one.
                     if (Aura * aura = target->GetAura(74396))
                     {
